Move tests to renamed newwindow package.
