+++ plugins/org.csstudio.openfile.newwindow.test/src/org/csstudio/openfile/newwindow/LinkUpdaterTest.java
@@ -1,4 +1,4 @@
-package org.csstudio.openfile.beamlines;
+package org.csstudio.openfile.newwindow;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
